services/backend/requirements.txt
fastapi==0.115.9
psycopg2-binary==2.9.9
websockets==14.1

SQLAlchemy==1.4.52
SQLAlchemy-Utils==0.33.0

alembic==1.14.0
pydantic==2.10.2
requests==2.32.3
starlette==0.41.3
uvicorn[standard]==0.32.0
python-multipart==0.0.6
openpyxl==3.1.5
openai==1.93.0
PyYAML==6.0.1
ruamel.yaml==0.18.6
typing-extensions>=4
pytest==8.3.2
boto3==1.36.26

google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.162.0

PyJWT==2.8.0
email-validator==2.1.0
